requirements.txt
numpy
pandas
matplotlib
seaborn
pillow
scipy
h5py
torch
torchvision
ultralytics
opencv-python
jupyter
